added control panel & download for referral
