+++ src/index.css
@@ -287,6 +287,10 @@ a.underlined-style {
   }
 }
 
+.preserve-whitepace {
+  white-space: pre-line;
+}
+
 /** Things only available on desktop */
 
 .display-desktop {
